requirements.txt
Django>=3.0,<4.0
google-api-python-client>=2.0.0
google-auth>=1.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
python-dotenv>=0.15.0